Move v2 utilities to v2 package
