+++ pkg/controller/v2/proposal/controller.go
@@ -14,8 +14,8 @@ import (
 	controllerutils "github.com/onosproject/onos-config/pkg/controller/utils"
 	proposalstore "github.com/onosproject/onos-config/pkg/store/v2/proposal"
 	pathutils "github.com/onosproject/onos-config/pkg/utils/path"
-	"github.com/onosproject/onos-config/pkg/utils/tree"
-	utilsv2 "github.com/onosproject/onos-config/pkg/utils/values/v2"
+	"github.com/onosproject/onos-config/pkg/utils/v2/tree"
+	utilsv2 "github.com/onosproject/onos-config/pkg/utils/v2/values"
 	"github.com/openconfig/gnmi/proto/gnmi_ext"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
