+++ pkg/northbound/admin/admin.go
@@ -17,8 +17,8 @@ import (
 	"github.com/onosproject/onos-config/pkg/store/v2/transaction"
 	"github.com/onosproject/onos-config/pkg/utils"
 	pathutils "github.com/onosproject/onos-config/pkg/utils/path"
-	"github.com/onosproject/onos-config/pkg/utils/tree"
-	valueutils "github.com/onosproject/onos-config/pkg/utils/values/v2"
+	"github.com/onosproject/onos-config/pkg/utils/v2/tree"
+	valueutils "github.com/onosproject/onos-config/pkg/utils/v2/values"
 	"github.com/onosproject/onos-lib-go/pkg/errors"
 	"github.com/onosproject/onos-lib-go/pkg/logging"
 	"github.com/onosproject/onos-lib-go/pkg/northbound"
