+++ pkg/northbound/gnmi/v2/set.go
@@ -24,7 +24,7 @@ import (
 
 	transactionstore "github.com/onosproject/onos-config/pkg/store/v2/transaction"
 	"github.com/onosproject/onos-config/pkg/utils"
-	valueutils "github.com/onosproject/onos-config/pkg/utils/values/v2"
+	valueutils "github.com/onosproject/onos-config/pkg/utils/v2/values"
 	"github.com/openconfig/gnmi/proto/gnmi"
 	"github.com/openconfig/gnmi/proto/gnmi_ext"
 )
